+++ Examples/MAX78000/TFT_Demo/src/utils.c
@@ -41,14 +41,14 @@
 /************************   STATIC FUNCTIONS  *******************************/
 
 /************************   PUBLIC FUNCTIONS  *******************************/
-uint32_t utils_get_time_ms(void)
+unsigned int utils_get_time_ms(void)
 {
     uint32_t sec, ssec;
     double subsec;
     uint32_t ms;
 
     MXC_RTC_GetSubSeconds(&ssec);
-    subsec = (double)ssec / 4096.0;
+    subsec = (double)ssec / (double)4096.0;
 
     MXC_RTC_GetSeconds(&sec);
