+++ Examples/MAX78002/CNN/pascalvoc-retinanetv7_3/src/cnn/nms.c
@@ -99,7 +99,7 @@ uint32_t utils_get_time_ms(void)
     uint32_t ms;
 
     MXC_RTC_GetSubSeconds(&ssec);
-    subsec = (double)ssec / 4096.0;
+    subsec = (double)ssec / (double)(4096.0);
 
     MXC_RTC_GetSeconds(&sec);
